Section8-GenAI/text-summarization-app/requirements.txt
google-cloud-aiplatform
Flask==2.1.0
gunicorn==20.1.0
python-docx